+++ src/shader/mod.rs
@@ -1,7 +1,9 @@
+pub mod constants;
 pub mod glsl;
 pub mod read;
 pub mod service;
 
+pub use constants::*;
 pub use glsl::*;
 pub use read::*;
 pub use service::*;
