+++ src/shader/read.rs
@@ -9,18 +9,7 @@ use std::path::{Path, PathBuf};
 
 use crate::shader::ShaderError;
 use crate::utils::{include_statement_from_string, pragma_shader_name};
-
-const USE_SKUGGBOX_CAMERA: &str = "#ifdef USE_SKUGGBOX_CAMERA
-    uniform mat4 sbCameraTransform;
-    void skuggbox_camera(vec2 uv, inout vec3 ro, inout vec3 rd) {
-        ro = sbCameraTransform[3].xyz;
-        rd = mat3(sbCameraTransform) * normalize(vec3(uv, 1));
-    }
-    #else
-    void skuggbox_camera(vec2 uv, inout vec3 ro, inout vec3 rd) {
-        // empty
-    }
-    #endif";
+use crate::SKUGGBOX_CAMERA;
 
 /// Read a shader from disk, return String or ShaderError
 pub fn read_shader_src(shader_path: PathBuf) -> anyhow::Result<String, ShaderError> {
@@ -108,26 +97,28 @@ impl PreProcessor {
 
     pub fn reload(&mut self) {
         match read_shader_src(self.main_shader_path.clone()) {
-            Ok(src) => self.main_shader_src = src,
-            Err(e) => {
-                error!("Could not re-compile shader {:?}", e)
-            }
+            Ok(src) => self.process(src),
+            Err(e) => error!("Could not re-compile shader {:?}", e),
         };
+    }
 
+    pub fn process(&mut self, shader_src: String) {
+        self.main_shader_src = shader_src;
         self.process_includes();
         self.process_integrations();
         self.recreate_file_list();
     }
 
-    pub fn process_pragma(&self, line: &str) -> Option<PragmaDirective> {
+    /// Handle pragma directives which are not
+    fn process_pragma(&self, line: &str) -> Option<PragmaDirective> {
         let camera_regex = Regex::new(r"^\s*#pragma\s+skuggbox\s*\(\s*camera\s*\)\s*$").unwrap();
 
         if line.contains("#pragma") && camera_regex.is_match(line) {
             debug!("Found camera integration: {:?}", line);
 
             let pragma_content = match self.use_camera_integration {
-                true => "#define USE_SKUGGBOX_CAMERA\n".to_string() + USE_SKUGGBOX_CAMERA,
-                _ => USE_SKUGGBOX_CAMERA.to_string(),
+                true => "#define USE_SKUGGBOX_CAMERA\n".to_string() + SKUGGBOX_CAMERA,
+                _ => SKUGGBOX_CAMERA.to_string(),
             };
             return Some(PragmaDirective::Camera(pragma_content));
         }
